Function for get dependencies recursively, excluding R base packages.

Function for get dependencies recursively, excluding R base packages.

Usage

search_dependencies(packages, type = "Imports")

Arguments

packages

vector with packages names

type

type of dependencies. Like package_dependencies from tools package.

Examples


options(repos = c(CRAN = 'https://cran.rstudio.com/'))
search_dependencies(c('dplyr', 'shiny', 'data.table'))
#>  [1] "R6"          "Rcpp"        "base64enc"   "bslib"       "cachem"     
#>  [6] "cli"         "commonmark"  "crayon"      "digest"      "fansi"      
#> [11] "fastmap"     "fontawesome" "fs"          "generics"    "glue"       
#> [16] "htmltools"   "httpuv"      "jquerylib"   "jsonlite"    "later"      
#> [21] "lifecycle"   "magrittr"    "memoise"     "mime"        "pillar"     
#> [26] "pkgconfig"   "promises"    "rappdirs"    "rlang"       "sass"       
#> [31] "sourcetools" "tibble"      "tidyselect"  "utf8"        "vctrs"      
#> [36] "withr"       "xtable"